// types/cep.ts
export interface ICep {
  cep?: string | undefined;
  logradouro?: string | undefined;
  complemento?: string | undefined;
  bairro?: string | undefined;
  localidade?: string | undefined;
  uf?: string | undefined;
  ibge?: string | undefined;
  gia?: string | undefined;
  ddd?: string | undefined;
  siafi?: string | undefined;
}